backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "televault"
version = "1.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.34,<0.35",
    "telethon>=1.36,<2.0",
    "cryptg>=0.4",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30,<0.31",
    "alembic>=1.13,<2.0",
    "pyjwt>=2.8,<3.0",
    "cryptography>=44.0,<45.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic-settings>=2.2,<3.0",
    "pillow>=11.0,<12.0",
    "streaming-form-data>=1.14,<2.0",
    "base58>=2.1,<3.0",
    "python-multipart>=0.0.9,<1.0",
]
